requirements-dev.txt
-r requirements.txt
django-webtest==1.8.0
factory-boy==2.8.1
freezegun==0.3.8
mkdocs==0.16.1
mock==2.0.0
pyquery==1.2.17
pytest==3.0.5
pytest-cov==2.4.0
pytest-django==3.1.2
pytest-flakes==1.0.1
pytest-mock==1.5.0
transifex-client==0.12.2
WebTest==2.0.24
